added admins to devise.

+++ app/controllers/admin/songs_controller.rb
@@ -7,14 +7,23 @@ class Admin::SongsController < ApplicationController
     @songs = Song.all
   end
   def new
+    unless current_admin
+      redirect_to :back, :alert => "Access denied."
+    end
    @song = Song.new()
   end
   def create
+    unless current_admin
+      redirect_to :back, :alert => "Access denied."
+    end
     song = Song.new(song_params)
     song.save!
     redirect_to(admin_songs_path)
   end
   def destroy
+    unless current_admin
+      redirect_to :back, :alert => "Access denied."
+    end
     song = Song.find(params[:id])
     song.destroy
     redirect_to admin_songs_path
